moved some kompoentens and files, and fixed comment
