+++ src/components/Lesson/InstructionButton.scss
@@ -1,0 +1,3 @@
+.instructionButton {
+  margin: 10px 10px 0 0;
+}
