+++ src/components/Lesson/Lesson.scss
@@ -20,5 +20,5 @@
 }
 
 .resetButton {
-  margin-right: 10px;
+  margin: 10px 10px 0 0;
 }
